Version 3.0.0a, a beta release. A significant refactor.
See the documentation changelog.

This is a beta release for version 3.0.0. There are a lot of changes
in this release and many of them impact the public APIs. The JavaDoc
has not yet been updated.

Includes all of the updates in CoffeeGrinder 3.0.0a.

Renamed the parser attributes used for pragmas to be URIs.

Removed the “rewrite” pragma; it can be replaced with standard
features: deletion and insertion. Removed the default priority pragma
because it didn’t actually seem useful.

Removed the notion of compiled grammars and the cache of compiled
grammars.

Refactored InvisibleXmlDocument to support the new tree construction
classes in CoffeeGrinder; added an InvisibleXmlFailureDocument
subclass to represent failed parses.

Added an option to omit the headers when parsing to CSV.

Lots of test improvements, rewrites.